Squashed commit of the following:
commit 5c20deafdb6c7a8a32d77ee3cbf9e3ac92f71377
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 14:59:59 2014 +0000

    Texture Management

commit d2ed1e5f95cff64098f19e5a37345a4d45644408
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 14:39:25 2014 +0000

    DrawSpriteState.cs

commit d3d9a64234dff4f1da443c3ff05fdf99c745da3d
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 13:57:55 2014 +0000

    Notes.

commit af94cdc7050bac2e3276c19b7ec6722251bed103
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 13:57:40 2014 +0000

    Simple TextureManager

commit eda1f93bd5b5b8ff0d8573cfd4a70723497c9e59
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 13:43:46 2014 +0000

    Tidying.

commit 798e0cc68cd47b3afbcd270e06189518f2b2c778
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 13:40:06 2014 +0000

    Marking regions for new classes to handle.

commit 2be9106d5361c5493f3f1450a2de76677eeac630
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 13:35:35 2014 +0000

    Added states.

commit 1fe8c583cb91dc0e0edde7cb9ca3601681e99023
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 13:35:14 2014 +0000

    Corrected states.

commit 4eb66ff44033d47828dc831d1d274a16a2864d05
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 13:34:39 2014 +0000

    Reinstated Input

commit cb69ec2989128bebc7bd25bcacad680c79186ebf
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 13:07:14 2014 +0000

    progress

commit 3e9b1b100dd8df5e0638c8313ade4f86dbdbefd5
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 13:07:02 2014 +0000

    Refactor.

commit 08d0414ce6d92acd26ccf2e5fd0ac56fc9cf1613
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 12:35:47 2014 +0000

    States and Tests!

commit ee817d345e4c9417cfa9cb943723fea843ec6705
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 04:09:15 2014 +0000

    Configurations changes!

commit 0a3ae393fc46bf259d07b5607dc58f1ec5588ee5
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 03:54:39 2014 +0000

    Testing!

commit 7f670873cd08ef3be48783cf48706935596d44e2
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 01:51:39 2014 +0000

    Timer again

commit 190275fbf7ce6d5dbdf35bc2a409bbe1ff3c5b6c
Author: Anthony Woodward <a.woodward6237@student.leedsmet.ac.uk>
Date:   Fri Jan 24 01:12:38 2014 +0000

    Removed experimental code.

+++ AWGL/Utilities/PreciseTimer.cs
@@ -25,11 +25,12 @@ namespace AWGL.Utilities
             GetElapsedTime();
         }
 
-        public double GetElapsedTime()
+        public float GetElapsedTime()
         {
             long time = 0;
             QueryPerformanceCounter(ref time);
-            double elapsedTime = (double)(time - m_previouslyElapsedTime) / (double)m_ticksPerSecond;
+
+            float elapsedTime = (float)(time - m_previouslyElapsedTime) / (float)m_ticksPerSecond;
 
             return elapsedTime;
         }
